requirements.txt
# Python Dependencies for OkCupid Automation
playwright>=1.40.0
websockets>=12.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.11
sqlalchemy>=2.0.23
asyncpg>=0.29.0
aiohttp>=3.9.1
Pillow>=10.2.0
numpy>=1.26.2

